backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
# For YOLO model
ultralytics==8.0.196
# For LLM interaction
openai==1.3.7
# For image processing
pillow==10.1.0
numpy==1.26.0 